refactor(__init__.py): move some functions into separate files
This moves toward using __init__.py files primarily for exposing public attributes.

I've tested a few examples (MultiFilter, SurfacePicking, and VTPViewer), and they seem to be
working with these changes.
