+++ trame/app/__init__.py
@@ -1,0 +1,8 @@
+from .app import activate_app, create_app, deactivate_app, get_app_instance
+
+__all__ = [
+    "activate_app",
+    "create_app",
+    "deactivate_app",
+    "get_app_instance",
+]
