+++ trame/utils/client_window_process.py
@@ -1,0 +1,41 @@
+from multiprocessing import Process
+
+
+class ClientWindowProcess(Process):
+    def __init__(
+        self, title=None, port=None, msg_queue=None, file_dialog=None, **kwargs
+    ):
+        Process.__init__(self)
+        self._title = title
+        self._port = port
+        self._msg_queue = msg_queue
+        self._window_args = kwargs
+        self._file_dialog = file_dialog
+        self._main_window = None
+
+    def _open_file_dialog(self):
+        result = self._main_window.create_file_dialog(**self._file_dialog)
+        self._msg_queue.put(["file_dialog", result])
+
+    def exit(self):
+        self._main_window.destroy()
+        self._msg_queue.put("closing")
+
+    def run(self):
+        try:
+            import webview
+        except:
+            print("trame.start_desktop_window() require pywebview==3.4")
+            return
+
+        self._main_window = webview.create_window(
+            title=self._title,
+            url=f"http://localhost:{self._port}/",
+            **self._window_args,
+        )
+        self._main_window.closing += self.exit
+
+        if self._file_dialog:
+            webview.start(func=self._open_file_dialog)
+        else:
+            webview.start()
