+++ trame/utils/logging.py
@@ -1,0 +1,2 @@
+def log_js_error(message):
+    print(f" > JS error | {message}")
